# backend/openapi.yml
openapi: 3.0.0
info:
  title: User Management API
  description: API for managing users
  version: 1.0.0
servers:
  - url: http://localhost:8081/api
paths:
  /users:
    get:
      summary: Get all users
      description: Retrieves a list of all users
      responses:
        "200":
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    delete:
      summary: Delete all users
      description: Deletes all users from the database
      responses:
        "204":
          description: All users deleted successfully
  /users/{id}:
    get:
      summary: Get a user by ID
      description: Retrieves a user by their ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
    put:
      summary: Update a user by ID
      description: Updates user details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
    delete:
      summary: Delete a user by ID
      description: Deletes a specific user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: User deleted successfully
        "404":
          description: User not found
  /createUser:
    post:
      summary: Create a new user
      description: Adds a new user to the database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        dv:
          type: string
        rut:
          type: string